pyproject.toml
[project]
name = "kn-fetch"
version = "1.0.0"
description = "全量代码资产AI扫描与语义建模系统"
authors = [
    {name = "KN-Fetch Team", email = "team@kn-fetch.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["code-analysis", "ai", "semantic-modeling", "refactoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    # 核心依赖
    "PyYAML>=6.0.1",
    "GitPython>=3.1.41",
    "tqdm>=4.66.1",
    "pandas>=2.1.4",
    "networkx>=3.2.1",
    "SQLAlchemy>=2.0.25",
    "pydantic>=2.5.3",
    "python-multipart>=0.0.6",
    
    # 代码解析
    "tree-sitter>=0.20.4",
    "tree-sitter-python>=0.20.4",
    "tree-sitter-javascript>=0.20.1",
    "tree-sitter-typescript>=0.20.3",
    "tree-sitter-java>=0.20.2",
    "tree-sitter-cpp>=0.20.2",
    "tree-sitter-go>=0.20.0",
    
    # LLM交互
    "openai>=1.10.0",
    "httpx>=0.26.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.20",
    "langgraph>=0.0.20",
    
    # 文档处理
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "markdown>=3.5.2",
    "Jinja2>=3.1.3",
    
    # 大规模处理
    "psutil>=5.9.8",
    "numpy>=1.26.3",
    
    # Web界面
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    
    # 测试
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    
    # 其他
    "python-dateutil>=2.8.2",
    "requests>=2.31.0",
    "chardet>=5.2.0",
    
    # design-v1新增依赖
    "loguru>=0.7.2",
    "pydantic-settings>=2.1.0",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "neo4j>=5.16.0",
    "pymilvus>=2.3.7",
    "lizard>=1.17.10",
    "semgrep>=1.43.0",
    "sonarqube>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.9",
    "black>=23.11.0",
    "isort>=5.13.2",
    "mypy>=1.7.1",
    "pre-commit>=3.5.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["B008", "B009", "B010"]

[tool.ruff.isort]
known-first-party = ["kn_fetch"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]